+++ NEWS.md
@@ -1,3 +1,14 @@
+# ontologyIndex 2.2
+
+* The `*_branches` functions have all been renamed to `*_descendants` and their `branch_roots` argument to `roots`.
+* The `get_descendants` function has had two arguments renamed: `ancestor` is now `roots` and can accept a vector of root terms, returning the set of all terms which descend from them; `remove_ancestor` is now `exclude_roots`, determining whether or not to exclude the given `roots` terms from the result.
+* A `check` function which checks whether an ontology is valid has been added.
+* XML encoded 'OWL' files can now be read in as `ontology_index`es.
+
+# ontologyIndex 2.1
+
+* A specialisation of `print` has been added for `ontology_index`es.
+
 # ontologyIndex 2.0
 
 * `get_ontology` now only requires a `file` argument.